Breast-Cancer-AI-App/requirements.txt
streamlit
torch
torchvision
Pillow
joblib
scikit-learn
numpy